create template parser
